% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_throw.R
\name{pin_throw}
\alias{pin_throw}
\title{Pin throw (pin_write equivalent)}
\usage{
pin_throw(
board,
file,
name,
comment = NULL,
...)
}
\arguments{
\item{board}{A pins board object from board_rsconnect()}

\item{file}{Object to be pinned}

\item{name}{Name of object to be pinned}

\item{comment}{A description of the pin to be pinned}

\item{...}{Any additional arguments for pins::pin_write}
}
\value{
A pinned object to the specified pins board
}
\description{
Writes a pin to to specified board while recording metadata to kingpin
}
\examples{
# Basic usage, assuming .Renviron is set up with CONNECT_SERVER and CONNECT_API_SERVER environmental variables:
library(kingpin)
board <- pins::board_rsconnect(server = Sys.getenv("CONNECT_SERVER"), key = Sys.getenv("CONNECT_API_KEY"))

kingpin::pin_throw(board, data.frame(a = 1:10, b = 1:10), "temp")

# To check if kingpin has updated:
kingpin <- pins::pin_read(board, "kingpin")$result$records

}
