% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_kingpin.R
\name{setup_kingpin}
\alias{setup_kingpin}
\title{Setup kingpin file to board}
\usage{
setup_kingpin(
server,
key,
force = FALSE)
}
\arguments{
\item{server}{URL of the board server. It's recommended to store these details in .Renviron and use Sys.getenv()}

\item{key}{API key to access the board. It's recommended to store these details in .Renviron and use Sys.getenv()}

\item{force}{Whether to force an override of an existing kingpin.}
}
\value{
A new pin named "kingpin" will be pinned on the board supplied to the function.
}
\description{
Sets up an empty "kingpin" pin to a board to be populated by activity data
}
\examples{
# Basic usage, assuming .Renviron is set up with CONNECT_SERVER and CONNECT_API_SERVER environmental variables:
setup_kingpin(server = Sys.getenv("CONNECT_SERVER"), key = Sys.getenv("CONNECT_API_KEY"))

}
